updatev2

+++ app/Http/Controllers/SettingController.php
@@ -9,7 +9,7 @@ use App\Models\User;
 use App\Models\UserOrganization;
 use App\Models\RegionUser;
 use App\Models\Sity;
-use App\Models\client;
+use App\Models\Client;
 use App\Models\Area;
 use App\Models\Organization;
 
@@ -20,23 +20,23 @@ class SettingController extends Controller
 
         $users = User::get();
 
-        foreach ( $users as $user) {
+        foreach ($users as $user) {
             $userOrg = UserOrganization::where('user_id', $user->id)->first();
-            if($userOrg) {
+            if ($userOrg) {
                 $user->organization_id = $userOrg->organization_id;
                 $user->role = $userOrg->role;
             }
 
             $regionUser = RegionUser::where('user_id', $user->id)->first();
-            if($regionUser) {
+            if ($regionUser) {
                 $user->areas = $regionUser->areas;
             }
 
             $user->save();
         }
-        
+
         dd('success');
-        
+
     }
 
     public function setting2()
@@ -44,18 +44,17 @@ class SettingController extends Controller
 
         $cities = Area::get();
 
-        foreach($cities as $city)
-        {
+        foreach ($cities as $city) {
             $org = Organization::where('id', $city->organization_id)->count();
             $reg = Sity::where('id', $city->city_id)->count();
 
-            if(!$org || !$reg) {
+            if (!$org || !$reg) {
                 $city->delete();
             }
         }
-        
+
         dd('success');
-        
+
     }
 
     public function setting3()
@@ -63,17 +62,16 @@ class SettingController extends Controller
 
         $cities = Sity::get();
 
-        foreach($cities as $city)
-        {
+        foreach ($cities as $city) {
             $org = Organization::where('id', $city->organization_id)->count();
 
-            if(!$org) {
+            if (!$org) {
                 $city->delete();
             }
         }
-        
+
         dd('success');
-        
+
     }
 
     public function setting4()
@@ -82,20 +80,19 @@ class SettingController extends Controller
         $clients = Client::get();
 
         $x = 0;
-        foreach($clients as $client)
-        {
+        foreach ($clients as $client) {
             $org = Organization::where('id', $client->organization_id)->count();
             $reg = Sity::where('id', $client->city_id)->count();
             $area = Area::where('id', $client->area_id)->count();
 
-            if(!$org || !$reg || !$area) {
+            if (!$org && !$reg && !$area) {
                 $client->delete();
-                $x ++;
+                $x++;
             }
         }
-        
+
         dd('success', $x);
-        
+
     }
 
 }
